Even more sanity

+++ code/game/gamemodes/meteor/meteors.dm
@@ -169,6 +169,7 @@
 		return ..() //Refer to atom/proc/Cross
 
 /obj/item/projectile/meteor/Bump(atom/A)
+
 	if(loc == null)
 		return
 
@@ -186,6 +187,9 @@
 
 /obj/item/projectile/meteor/radioactive/Bump(atom/a)
 
+	if(loc == null)
+		return
+
 	for(var/mob/living/M in viewers(src, null))
 		M.radiation += rand(5, 10)
 
@@ -198,6 +202,8 @@
 	pass_flags = PASSTABLE
 
 /obj/item/projectile/meteor/small/Bump(atom/A)
+	if(loc == null)
+		return
 
 	explosion(get_turf(src), -1, 1, 3, 4, 0, 1, 0) //Tiny meteor doesn't cause too much damage
 	qdel(src)
@@ -208,6 +214,9 @@
 	icon_state = "small_flash"
 
 /obj/item/projectile/meteor/small/flash/Bump(atom/A)
+
+	if(loc == null)
+		return
 
 	//Adjusted from flashbangs, should be its own global proc
 	visible_message("<span class='danger'>BANG</span>")
@@ -256,7 +265,8 @@
 				to_chat(M, "<span class='warning'>You can't hear anything!</span>")
 				M.sdisabilities |= DEAF
 
-	..()
+	explosion(get_turf(src), -1, 1, 3, 4, 0, 1, 0) //Tiny meteor doesn't cause too much damage
+	qdel(src)
 
 /obj/item/projectile/meteor/piercing
 	name = "piercing meteor"
@@ -265,6 +275,9 @@
 	var/pierce_health = 1 //When 0, piercing meteor explodes like normal
 
 /obj/item/projectile/meteor/piercing/Bump(atom/A)
+
+	if(loc == null)
+		return
 
 	if(pierce_health)
 		explosion(get_turf(A), 1, 0, 0, 0, 0, 1, 0) //Blow up the resisting object
@@ -280,6 +293,9 @@
 
 /obj/item/projectile/meteor/big/Bump(atom/A)
 
+	if(loc == null)
+		return
+
 	explosion(get_turf(src), 4, 6, 8, 8, 0, 1, 0) //You have been visited by the nuclear meteor
 	qdel(src)
 
@@ -289,6 +305,9 @@
 	icon_state = "big_cluster"
 
 /obj/item/projectile/meteor/big/cluster/Bump(atom/A)
+
+	if(loc == null)
+		return
 
 	explosion(get_turf(A), 1, 0, 0, 0, 0, 1, 0) //Enough to destroy whatever was in the way
 	for(var/i = 0, i < 3, i++)
@@ -319,5 +338,9 @@
 	icon_state = "human"
 
 /obj/effect/meteor/gib/Bump(atom/A)
+
+	if(loc == null)
+		return
+
 	new /obj/effect/gibspawner/human(src.loc)
 	qdel(src)
